Enhance rental management and admin interfaces
- Updated token bucket configuration in arcjet.js to improve rate limiting.
- Refactored AdminRentals, AdminUserManagement, AdminVehicles, and Home components to utilize new shared components for forms, tables, and filters, enhancing code reusability and maintainability.
- Integrated pagination and modal components for better user experience in AdminUserManagement and AdminVehicles.
- Improved vehicle and user management functionalities with new filtering options and streamlined data handling.
- Enhanced RentalCheckout to prevent duplicate processing during payment handling.
