+++ backend/lib/arcjet.js
@@ -16,9 +16,9 @@ export const aj = arcjet({
     }), //detects bots based on behavior and characteristics. Doesnt block search engines https://arcjet.com/bot-list
     tokenBucket({
       mode: "LIVE",
-      refillRate: 5, //number of tokens to add per interval (5 every 10 seconds = 30 per minute with a max of 10 at any time)
-      interval: 10, //interval in seconds
-      capacity: 10, //max number of tokens in the bucket
+      refillRate: 10, //number of tokens to add per interval (5 every 10 seconds = 30 per minute with a max of 10 at any time)
+      interval: 5, //interval in seconds
+      capacity: 30, //max number of tokens in the bucket
     }),
   ],
 });
