+++ frontend/src/components/shared/ConfirmModal.jsx
@@ -1,0 +1,21 @@
+function ConfirmModal({ title, message, onCancel, onConfirm }) {
+  return (
+    <div className="modal-overlay">
+      <div className="modal-content">
+        <h3>{title}</h3>
+        <p>{message}</p>
+        <div className="modal-actions">
+          <button className="secondary-btn" onClick={onCancel}>
+            Cancel
+          </button>
+          <button className="danger-btn" onClick={onConfirm}>
+            Confirm
+          </button>
+        </div>
+      </div>
+    </div>
+  );
+}
+
+export default ConfirmModal;
+
