+++ frontend/src/utils/dateUtils.js
@@ -1,0 +1,16 @@
+/**
+ * Formats a date string to a human-readable format
+ * @param {string} dateStr - ISO date string
+ * @returns {string} Formatted date string
+ */
+export const formatDate = (dateStr) => {
+  if (!dateStr) return "";
+  const d = new Date(dateStr);
+  if (Number.isNaN(d.getTime())) return dateStr;
+  return d.toLocaleDateString(undefined, {
+    year: "numeric",
+    month: "short",
+    day: "numeric",
+  });
+};
+
